pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "assgen"
dynamic = ["version"]
description = "Game asset generation CLI — client/server pipeline for AI-driven 3D game asset production"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "typer[all]>=0.12",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "platformdirs>=4.2",
    "pyyaml>=6.0",
    "rich>=13.7",
    "huggingface-hub>=0.23",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "anyio>=4.4",
]

[project.optional-dependencies]
inference = [
    "transformers>=4.40",
    "torch>=2.3",
    "diffusers>=0.28",
    "accelerate>=0.30",
    "trimesh>=3.21",
    "Pillow>=10.0",
]
spaces = [
    # Core inference
    "transformers>=4.40",
    "torch>=2.3",
    "diffusers>=0.28",
    "accelerate>=0.30",
    "trimesh>=3.21",
    "Pillow>=10.0",
    # Audio processing
    "pydub>=0.25",
    "pyloudnorm>=0.1",
    "scipy>=1.13",
    # Visualization
    "matplotlib>=3.8",
    # Mesh utilities
    "pyfqmr>=0.2",
    "xatlas>=0.0.9",
    # Voice cloning
    "TTS>=0.22",
    # Upscaling (requires opencv)
    "basicsr>=1.4.2",
    "realesrgan>=0.3.0",
    "opencv-python-headless>=4.9",
]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "ruff>=0.4",
]
docs = [
    "mkdocs-material>=9.5",
    "mkdocs-typer>=0.0.2",
    "mkdocstrings[python]>=0.25",
    "mkdocs-gen-files>=0.5",
    "griffe>=0.45",
]

[project.scripts]
assgen = "assgen.client.cli:app"
assgen-server = "assgen.server.cli:app"

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/assgen/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/assgen"]

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in default arguments
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]  # allow assert in tests

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "integration: tests that make real network calls (deselect with -m 'not integration')",
    "slow: tests that run real model inference (skip with -m 'not slow')",
    "gpu: tests that require a CUDA GPU",
]
addopts = "--tb=short"

[tool.coverage.run]
source = ["src/assgen"]
omit = ["*/tests/*", "*/_version.py"]

[tool.coverage.report]
show_missing = true
skip_covered = false
